% Generated by roxygen2 (4.0.0): do not edit by hand
\name{export_data}
\alias{export_data}
\title{Convenience function to export data from multiple
IrmsData objects of the same class into a comma-separated value file.}
\usage{
export_data(data, file = "irms_data_export.csv", ...)
}
\description{
Convenience function to export data from multiple
IrmsData objects of the same class into a comma-separated value file.
}

