% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ascii}
\alias{get_ascii}
\title{Get ASCII text from raw binary data}
\usage{
get_ascii(rawdata, minlength = 1, value = TRUE, all = FALSE)
}
\arguments{
\item{value}{if FALSE, returns pattern indices instead}

\item{all}{if TRUE, returns vector of all found patterns}
}
\value{
vector of value(s) or start indices, NULL if none foud
}
\description{
Get ASCII text from raw binary data
}

