% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_raw_binary_data}
\alias{get_raw_binary_data}
\title{Get raw binary data in text form}
\usage{
get_raw_binary_data(rawdata)
}
\value{
character string version of raw binary data
}
\description{
Get raw binary data in text form
}

