\name{IDP.readIsodatFile}
\alias{IDP.readIsodatFile}
\title{read a standard isodat file with H2 data in it
printRemoved - if you want command line output of keys that are removed in the cleanup process}
\usage{
IDP.readIsodatFile(file, printRemoved = FALSE)
}
\description{
read a standard isodat file with H2 data in it printRemoved
- if you want command line output of keys that are removed
in the cleanup process
}

