% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IrmsContinuousFlowDataClass.R
\docType{class}
\name{IrmsContinuousFlowData}
\alias{IrmsContinousFlowData}
\alias{IrmsContinuousFlowData}
\alias{isoplot}
\title{IrmsContinuousFlowData reference class}
\arguments{
\item{masses}{which masses to retrieve, all defined ones by defaul}

\item{melt}{whether to melt the data frame}

\item{ratios}{which ratios to retrieve, all defined ones by default}

\item{melt}{whether to melt the data frame}

\item{rts}{Get peaks by name}

\item{rts}{Set peak(s) columns (this one should not be used externally)}

\item{peak_nr}{- peak number(s) to update with the provided attributes}

\item{...}{- peak columns to change (only one value per attribute allowed!)
set peak(s) columns by retention time
Set/unset reference peaks

Identify peaks as reference peaks (or remove their status as a reference peak).
This is a specialized call for \code{set_peak_by_rt}}

\item{rt}{(can be a vector)}

\item{set}{- wether to set it to be or not be a reference peak}

\item{reevalute}{- whether to revaluate the peak table right away
Identify peak(s)

Identify peaks by mapping compound names to retention times}

\item{rts}{- retention times}

\item{compounds}{- compound names
Map peak(s)

Add information to peaks by mapping properties from a data frame
that contains at least the defined \code{dataTableKeys['peak_nr']} or 
\code{dataTableKeys['rt']} as a column. Additional columns (other than peak
nr and retention time) are mapped
to the relevant peaks if they correspond to existing columns, otherwise
they are disregarded with a warning.}

\item{map}{- the map of properties
Evaluate data in peak table
Plot the data points in the peak table}

\item{y}{= expression which data to plot (will be evaluated in context of the data frame)}

\item{ylab}{= y axis label}

\item{title}{= title of the plot}

\item{data}{= peak table data (by default the whole peak table)}

\item{y}{= expression which data to plot (will be evaluated in context of the data frame)}

\item{ylab}{= y axis label}

\item{title}{= title of the plot}

\item{data}{= peak table data (by default the whole peak table)
Make a ggplot of the references in the peak table
Make a ggplot of the data peaks in the peak table
Plot the data (both masses and ratios) - much faster than ggplot but not as versatile}

\item{tlim}{time range, should be in the same tunits}

\item{masses}{which masses to plot (all defined in plot optinos by default)}

\item{ratios}{which ratios to plot (all defined in plot options by default)}

\item{tunits}{time units, as defined in tunits (currently either "s" or "min"), takes the one set in plotOptions as default}

\item{tlim}{time range, should be in the same tunits}

\item{masses}{which masses to plot (all defined in plot optinos by default)}

\item{ratios}{which ratios to plot (all defined in plot options by default)}

\item{tunits}{time units, as defined in tunits (currently either 's' or 'min'), takes the one set in plotOptions as default
Plot the masses (this if much faster than ggplot but not as versatile)}

\item{tlim}{-> time range, should be in the same tunits}

\item{tunits}{-> time units, as defined in tunits (currently either "s" or "min"), takes the one set in plotOptions as default
Plot the ratios (this if much faster than ggplot but not as versatile)}

\item{tlim}{-> time range, should be in the same tunits}

\item{tunits}{-> time units, as defined in tunits (currently either "s" or "min"), takes the one set in plotOptions as default
ggplot the data}

\item{tlim}{time range (in tunits units)}

\item{tunits}{units (currently "s" or "min")}

\item{masses}{vector of the masses to plot (if NULL, panel excluded)}

\item{ratios}{vector of the ratios to plot (if NULL, panel excluded)}

\item{tlim}{time range (in tunits units)}

\item{tunits}{units (currently 's' or 'min')}

\item{masses}{vector of the masses to plot (if NULL, panel excluded)}

\item{ratios}{vector of the ratios to plot (if NULL, panel excluded)
summarize data to pdf}

\item{file}{the file name where to save, by default saves where the file was original from}

\item{whether}{to try to compact the pdf to make it smaller}

\item{plots}{whether to include the plot}

\item{table}{whether to include the table
export data (by default to csv)}
}
\value{
returns the data frame of found peaks (0-row df if none found)
Get peaks by rt
}
\description{
IrmsContinuousFlowData reference class
}
\section{Fields}{

\describe{
\item{\code{massData}}{stores the chromatographic data (the actual mass and ratio data traces), 
inherited from IrmsDataClass}

\item{\code{dataTable}}{stores the peak table (detected peaks and all their information),
inherited from IrmsDataClass}

\item{\code{dataTableKeys}}{stores information about which columns correspond
to key elements of the dataTable (e.g. the peak number, retention time
and compound name)}
}}
\section{Methods}{

\describe{
\item{\code{export_data(file, ...)}}{export the data stored in this object to file}

\item{\code{get_data_table(...)}}{retrieve the data table}

\item{\code{get_mass_data(masses = names(.self$plotOptions$masses), melt = FALSE, ...)}}{get the mass trace data for specific masses, can be provided in \code{melt = TRUE} long format
for easy use in ggplot style plotting}

\item{\code{get_peak(peak_nr, select = names(dataTable))}}{retrieve information for a peak in the peak table (identified by peak_nr), can specify which columns to retrieve
with \code{selec}, retrieves all columns by default}

\item{\code{get_peak_by_name(names, select = names(dataTable))}}{retrieve information for peak(s) in the peak table (identified by names)}

\item{\code{get_peak_by_rt(rts, select = names(dataTable))}}{retrieve information for peak(s) in the peak table (identified by retention times)}

\item{\code{get_peak_nr_by_name(names)}}{find peak numbers (i.e. ids) by name(s), returns a vector of found peak numbers (integer(0) if none found)}

\item{\code{get_peak_nr_by_rt(rts)}}{find peak numbers (i.e. ids) by retention time(s), returns a vector of found peak numbers (integer(0) if none found)}

\item{\code{get_ratio_data(ratios = names(.self$plotOptions$ratios), melt = FALSE)}}{get the ratio trace data for specific ratios, can be provided in \code{melt = TRUE} format
for easy use in ggplot style plotting}

\item{\code{identify_peaks(rts, compounds)}}{Identify peaks by mapping compound names to retention times}

\item{\code{init_irms_data()}}{initialize irms data container}

\item{\code{make_ggplot(...)}}{generate a ggplot object for the data in this IrmsData object}

\item{\code{map_peaks(map)}}{Add information to peaks by mapping properties from a data frame
that contains at least the defined peak number (e.g. `Peak Nr.`) or 
retention time (Rt) as a column. Additional columns (other than peak
nr and retention time) are mapped to the relevant peaks if they correspond 
to existing columns, otherwise they are disregarded with a warning. 

Note: make sure to have the data.frame that is passed in set with
\code{stringsAsFactors = F} (usually the desired setting for the mapping)}

\item{\code{plot(...)}}{plot data with standard plot functions (fast) to standard output}

\item{\code{plot_data(y, ylab = "", title = "data peaks")}}{plot the data of the actual sample peaks, see \code{plot_data_table} for details on syntax}

\item{\code{plot_data_table(y = NULL, ylab = "", title = "",
  data = get_data_table())}}{Plot the data points in the peak table

#' @param y = expression which data to plot (will be evaluated in context of the data frame)

#' @param ylab = y axis label
 
#' @param title = title of the plot
 
#' @param data = peak table data (by default the whole peak table)}

\item{\code{plot_masses(tlim = NULL, ylim = NULL,
  masses = names(.self$plotOptions$masses),
  tunits = .self$plotOptions$tunits$labels[[.self$plotOptions$tunits$value]])}}{Plot the masses (this if much faster than ggplot but not as versatile)}

\item{\code{plot_ratios(tlim = NULL, ylim = NULL,
  ratios = names(.self$plotOptions$ratios),
  tunits = .self$plotOptions$tunits$labels[[.self$plotOptions$tunits$value]])}}{Plot the ratios (this if much faster than ggplot but not as versatile)}

\item{\code{reevaluate_peak_table()}}{reevalutes the peak table (not currently implemented)}

\item{\code{set_ref_peaks(rts, set = TRUE, reevaluate = FALSE)}}{Identify peaks (by their retention times) as reference peaks (or remove their status as a reference peak)}

\item{\code{summarize(file, ....)}}{summarize the data stored in this object and save it to file}
}}
\note{
consider storing the extra x units in the massData but worries about the size of this
object keep me from storing extra information in massData
get data for ratios

make sure to have the data.frame that is passed in with
stringsAsFactors = F

consider adding a ... option to allow setting plot options
or temporarily overwriting them here
}
\seealso{
\link{IrmsData}, \link{IrmsDualInletData}
get data for masses
}

