% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{IrmsData-class}
\alias{IrmsData}
\alias{IrmsData-class}
\title{IrmsData reference class}
\description{
IrmsData reference class
}
\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}
}}
\note{
not sure yet what's best to abstract into this one
}

