% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{IrmsDualInletData}
\alias{IrmsDualInletData}
\title{IrmsDualInletData reference class}
\description{
IrmsDualInletData reference class
}
\note{
not implemented yet for any actual data reading
}
\seealso{
\link{IrmsData}, \link{IrmsContinuousFlowData}
}

