% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/IsodatClumpedCO2FileClass.R
\docType{class}
\name{IsodatClumpedCO2File}
\alias{IsodatClumpedCO2File}
\title{Clumped dual inlet data class}
\description{
Clumped dual inlet data class
}
\seealso{
\link{IrmsDualInletData}, \link{BinaryFile}, \link{IsodatFile}, \link{IsodatDualInletFile}
}

