% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/IsodatDualInletFileClass.R
\docType{class}
\name{IsodatDualInletFile}
\alias{IsodatDualInletFile}
\title{Clumped dual inlet data class}
\description{
Clumped dual inlet data class
}
\section{Methods}{

\describe{
\item{\code{export_data(file, ...)}}{export the data stored in this object to file}

\item{\code{initialize(file, ...)}}{initialize BinaryFile object, requires a file path}

\item{\code{process(...)}}{process the raw data to fill the data list}
}}
\seealso{
\link{BinaryFile}, \link{IsodatFile}, \link{IrmsDualInletData}, \link{IrmsData}
}

