% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsodatFileClass.R
\docType{class}
\name{IsodatFile}
\alias{IsodatFile}
\title{Isodat file class}
\description{
Class representing an isodat binary file.
}
\section{Methods}{

\describe{
\item{\code{extract_measurement_info()}}{extracts the measurement information}

\item{\code{extract_sequence_line_info()}}{extracts the sequence line information}
}}
\seealso{
\link{BinaryFile}
}

