% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsodatHydrogenContinuousFlowFileClass.R
\docType{class}
\name{IsodatHydrogenContinuousFlowFile}
\alias{IsodatHydrogenContinuousFlowFile}
\title{H-CSIA DataClass}
\arguments{
\item{read_mass_data}{whether to read chromatographic mass+ratio data (can be a lot of data)
Evaluate data in peak table

This function uses the standards defined in the peak table to (re)evaluate the isotopic
composition of the other peaks in this reference frame.

Make a ggplot of the references

By default, this makes a plot that generates an overview
of how close the reference peaks are to
each other.
Make a ggplot of the data

By default, this makes a plot of the delta values.}
}
\description{
Objects of this class hold the isotopic data from compound specific hydrogen
isotope analysis recorded in Isodat file formats (currently supported isodat
version is 2.0 for chromatographic and peak table data and isodat version 2.5 and
3.0 for chromatographic data only).
}
\details{
This class is derived indirectly from \link{IrmsContinuousFlowData} which defines a 
number of useful plotting, export and data access methods.
}
\section{Methods}{

\describe{
\item{\code{load(...)}}{load the data from the file and generate key lookup}

\item{\code{plot_data(y, ylab = "", title = "data peaks")}}{plot the data of the actual sample peaks, see \code{plot_data_table} for details on syntax}

\item{\code{reevaluate_peak_table()}}{reevalutes the peak table (not currently implemented)}
}}
\seealso{
\link{BinaryFile}, \link{IsodatFile}, \link{IsodatContinuousFlowFile}, \link{IrmsContinuousFlowData}, \link{IrmsData}
initialize irms data container
expand parent load function with key_cleanup
expand parent procdess function specifically for hydrogen continuous flow data
NOTE: this does NOT callSuper because the hydrogen file (.cf) seems very different from
generic IsodatContinuousFlowFile (.dxf)
}

