% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsodatScanFileClass.R
\docType{class}
\name{IsodatScanFile}
\alias{IsodatScanFile}
\title{Scan data class}
\description{
Scan data class
}
\section{Methods}{

\describe{
\item{\code{initialize(file, ...)}}{initialize BinaryFile object, requires a file path}

\item{\code{process(...)}}{process the raw data to fill the data list}
}}
\seealso{
\link{BinaryFile}, \link{IsodatFile}, \link{IrmsScanData}, \link{IrmsData}
initialize
initialize irms data container
expand process function specifically for dual inlet type data
custom show function to display roughly what data we've got going
}

