% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{isoplot}
\alias{isoplot}
\title{Generic plotting function for iso objects}
\usage{
isoplot(iso, type = c("gg", "interactive"), ...)
}
\description{
Calls the make_ggplot function internally
}
\note{
very basic implement, @SK fix this
}

