% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\docType{package}
\name{isoread-package}
\alias{isoread-package}
\title{isoread package}
\description{
R interface to IRMS (isotope ratio mass spectrometry) file formats typically used in stable isotope geochemistry.
}
\details{
See \code{\link{isoread}} for details on how to use.
}
\author{
Sebastian Kopf

Max Lloyd
}

