% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/isoread.R
\name{isoread}
\alias{isoread}
\title{Read isotope data files}
\usage{
isoread(files, type, load_chroms = T, ...)
}
\arguments{
\item{type}{type of the files to be read
\itemize{
   \item{'H_CSIA'}{ = compound specific IRMS data for hydrogen isotopes}
}}

\item{load_chroms}{whether to keep the chromatograms in the objects (otherwise only peak tables are kept)}

\item{...}{parameters passed to the \code{load} and \code{process} functions of the IsodatFile objects}

\item{file}{path to the file(s) to read}
}
\value{
List of file \code{type} specific objects.
\itemize{
   \item{'DI'}{ = instance(s) of a basic \code{\link{IsodatDualInletFile}} which implements \code{\link{IrmsDualInletData}}}.
   \item{'CO2_CLUMPED'}{ = instance(s) of the more specialized \code{\link{IsodatClumbedCO2File}} which extends \code{\link{IsodatDualInletFile}}}.
   \item{'CFLOW'}{ = instance(s) of a basic \code{\link{IsodatContinuousFlowFile}} which extends \code{\link{IsodatContinuousFlowData}}}.
   \item{'H_CSIA'}{ = instance(s) of \code{\link{IsodatHydrogenContinuousFlowFile}} which extends \code{\link{IsodatContinuousFlowFile}}}.
}
If file names start with a number,
then the number is used as key in the list, otherwise the whole filename is the key.
If there is only one file, the object is returned directly.
}
\description{
Reads isodat file(s) and returns the contents as file type specific
instances of \code{\link{BinaryFile}} and \code{\link{IrmsData}} (extends both).
}

