\name{isoread_folder}
\alias{isoread_folder}
\title{Reads all isodat files in a folder.
See \code{\link{isoread}} for paramter and return value details.}
\usage{
isoread_folder(folder, type, extension = ".cf", ...)
}
\description{
Reads all isodat files in a folder. See
\code{\link{isoread}} for paramter and return value
details.
}

