% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utilities.R
\name{map_binary_data_type}
\alias{map_binary_data_type}
\title{Binary data type mapping}
\usage{
map_binary_data_type(type = c("binary", "UTF8", "UTF16", "UTF32", "short",
  "long", "long long", "float", "double"))
}
\arguments{
\item{type}{\itemize{
   \item{'binary'}{ = raw with 1 byte (raw data)}
   \item{'UTF8'}{ = character with 1 byte (ascii)}
   \item{'UTF16'}{ = character with 2 bytes (unicode)}
   \item{'UTF32'}{ = character with 4 bytes (unicode)}
   \item{'short'}{ = integer with 2 bytes (16bit)}
   \item{'long'}{ = integer with 4 bytes (32bit)}
   \item{'longlong'}{ = integer with 8 bytes (64bit)}
   \item{'float'}{ = numeric with 4 bytes (32bit)}
   \item{'double'}{ = numeric with 8 bytes (64bit)}
}}
}
\description{
Maps binary C data types to proper R data types and byte lengths
}
\note{
implemented signed int and complex if needed
}

