% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{map_peaks}
\alias{map_peaks}
\title{Map peak table}
\usage{
map_peaks(iso, map, startRow = 3, libfile = NULL,
  colClasses = c("numeric", "text", "text"), ...)
}
\arguments{
\item{iso}{IrmsContinuousFlowData object(s)}

\item{map}{either a data frame with a map (containing column Rt and Component)
or the file path to a mapping file, extension determines how it will be processed
currently only xlsx and xls are supported. Excel file must include column headers
on the indicated row (default startRow = 3) with columns Rt and Component Component}

\item{libfile}{name of the library file, also only xlsx and xls currently supported
if provided, will attempt to merge the components in the mapping file with the library
information for additional details on Formula and other compound properties}
}
\description{
Map peak table data of IrmsContinuousFlowData object(s) based on a data
frame or input excel file.
}

