% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_binary_data}
\alias{parse_binary_data}
\title{Wrapper for parsing binary data.}
\usage{
parse_binary_data(data, type, length = 1)
}
\arguments{
\item{type}{data type see \code{\link{map_binary_data_type}} for details}

\item{length}{how many instances of this object (for characters and raw this means length of string, all others a vector)}
}
\value{
read data
}
\description{
Convenience wrapper for parsing binary data.
For more details on reading binary data, check ?readBin
}

