% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/isoread.R
\name{quickview}
\alias{quickview}
\title{File quickview}
\usage{
quickview(iso, reload = FALSE, show = c("Peak Nr.", "Status", "Ref. Peak",
  "Component", "Rt", "Start", "End", "Ampl. 2", "d 2H/1H"))
}
\arguments{
\item{iso}{a single isodat file obj}

\item{reload}{whether to reload the file (this is forced if there is no chromatographic data)}

\item{show}{list of peak table columns to show}
}
\description{
This functions serves to gain a quick view of a loaded isodat file.
It shows the masses plot and prints a minimal subset of the peak table. Optionally reloads
the file (tries to keep the peak definitions).
}

