% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/isoread.R
\name{reload}
\alias{reload}
\title{Reload an isodat file object.}
\usage{
reload(iso, remap_peaks = TRUE, load_chroms = TRUE)
}
\arguments{
\item{iso}{the object to reload (can be a list)}

\item{remap_peaks}{whether to keep the peak identification or not}

\item{load_chroms}{whether to load the chroms (much smaller object without)}
}
\value{
the reloaded obj (or list of objs)
}
\description{
Reload an existing isodat object with all the chromatographic data and
resets the peak table if keep_peaks = FALSE. Good for interrogation of
an individual file. Requires the original file to still be in the same
location.
}
\note{
currently only for type = "H_CSIA"
}

