% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/isoread.R
\name{summarize_all}
\alias{summarize_all}
\title{Summarize a collection of IrmsData objects}
\usage{
summarize_all(iso, ...)
}
\arguments{
\item{iso}{IrmsData object(s)}

\item{...}{all passed to summarize}
}
\description{
Summarize a collection of IrmsData objects
}

