% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{BinaryFile}
\alias{BinaryFile}
\title{Binary File reference class}
\description{
Binary File reference class
}
\section{Fields}{

\describe{
\item{\code{filepath}}{stores the path to the binart file}

\item{\code{filename}}{stores the filename}

\item{\code{creation_date}}{stores the date the file was created (if it could be retrieved,
which is not always the case when running on linux but no problem on OS X and windows)}

\item{\code{rawdata}}{this is the binary i}
}}

