% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{IrmsContinuousFlowData-class}
\alias{IrmsContinousFlowData}
\alias{IrmsContinuousFlowData-class}
\title{IrmsContinuousFlowData reference class}
\description{
IrmsContinuousFlowData reference class
}
\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}
}}

