% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{IrmsContinuousFlowData}
\alias{IrmsContinousFlowData}
\alias{IrmsContinuousFlowData}
\title{IrmsContinuousFlowData reference class}
\description{
IrmsContinuousFlowData reference class
}
\section{Fields}{

\describe{
\item{\code{chromData}}{stores the chromatographic data (the actual mass and ratio data traces),}

\item{\code{peakTable}}{stores the peak table (detected peaks and all their information)}

\item{\code{peakTableColumns}}{stores the definition of which columns exist in the
peak table and what their proper data types are}

\item{\code{peakTableKeys}}{stores information about which columns correspond
to key elements of the peakTable (e.g. the peak number, retention time
and compound name)}
}}
\section{Methods}{

\describe{
\item{\code{check_chrom_data(masses = names(.self$plotOptions$masses),
  ratios = names(.self$plotOptions$ratios), ..., warn = TRUE)}}{checks the consistency of the chromatographic data, by default checks for 
all masses and ratios}

\item{\code{check_peak_table(..., warn = TRUE)}}{checks the consistency of the peak table and converts data types if necessary}

\item{\code{export_data(file, ...)}}{export the data stored in this object to file}

\item{\code{get_mass_data(masses = names(.self$plotOptions$masses), melt = FALSE)}}{get the mass trace data for specific masses, can be provided in \code{melt = TRUE} format
for easy use in ggplot style plotting}

\item{\code{get_peak(peak_nr, select = names(peakTable))}}{retrieve information for a peak in the peak table (identified by peak_nr), can specify which columns to retrieve
with \code{selec}, retrieves all columns by default}

\item{\code{get_peak_by_name(names, select = names(peakTable))}}{retrieve information for peak(s) in the peak table (identified by names)}

\item{\code{get_peak_by_rt(rts, select = names(peakTable))}}{retrieve information for peak(s) in the peak table (identified by retention times)}

\item{\code{get_peak_nr_by_name(names)}}{find peak numbers (i.e. ids) by name(s), returns a vector of found peak numbers (integer(0) if none found)}

\item{\code{get_peak_nr_by_rt(rts)}}{find peak numbers (i.e. ids) by retention time(s), returns a vector of found peak numbers (integer(0) if none found)}

\item{\code{get_peak_table(type = c("ref", "data", "both"))}}{retrieve the peak table}

\item{\code{get_ratio_data(ratios = names(.self$plotOptions$ratios), melt = FALSE)}}{get the ratio trace data for specific ratios, can be provided in \code{melt = TRUE} format
for easy use in ggplot style plotting}

\item{\code{ggplot(...)}}{generate a ggplot object for the data in this IrmsData object}

\item{\code{identify_peaks(rts, compounds)}}{Identify peaks by mapping compound names to retention times}

\item{\code{init_irms_data()}}{initialize irms data container}

\item{\code{map_peaks(map)}}{Add information to peaks by mapping properties from a data frame
that contains at least the defined peak number (e.g. `Peak Nr.`) or 
retention time (Rt) as a column. Additional columns (other than peak
nr and retention time) are mapped to the relevant peaks if they correspond 
to existing columns, otherwise they are disregarded with a warning. 

Note: make sure to have the data.frame that is passed in set with
\code{stringsAsFactors = F} (usually the desired setting for the mapping)}

\item{\code{plot_data(y, ylab = "", title = "data peaks")}}{plot the data of the actual sample peaks, see \code{plot_peak_table} for details on syntax}

\item{\code{plot_masses(tlim = NULL, ylim = NULL,
  masses = names(.self$plotOptions$masses),
  tunits = .self$plotOptions$tunits$labels[[.self$plotOptions$tunits$value]])}}{Plot the masses (this if much faster than ggplot but not as versatile)}

\item{\code{plot_peak_table(y = NULL, ylab = "", title = "",
  data = get_peak_table())}}{Plot the data points in the peak table

#' @param y = expression which data to plot (will be evaluated in context of the data frame)

#' @param ylab = y axis label
 
#' @param title = title of the plot
 
#' @param data = peak table data (by default the whole peak table)}

\item{\code{plot_ratios(tlim = NULL, ylim = NULL,
  ratios = names(.self$plotOptions$ratios),
  tunits = .self$plotOptions$tunits$labels[[.self$plotOptions$tunits$value]])}}{Plot the ratios (this if much faster than ggplot but not as versatile)}

\item{\code{plot_refs(y, ylab = "", title = "references")}}{plot the data of the reference peaks, see \code{plot_peak_table} for details on syntax}

\item{\code{reevaluate_peak_table()}}{reevalutes the peak table (not currently implemented)}

\item{\code{set_ref_peaks(rts, set = TRUE, reevaluate = FALSE)}}{Identify peaks (by their retention times) as reference peaks (or remove their status as a reference peak)}

\item{\code{summarize(file, ....)}}{summarize the data stored in this object and save it to file}
}}
\seealso{
\link{IrmsData}, \link{IrmsDualInletData}
}

