% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{IrmsDualInletData-class}
\alias{IrmsDualInletData}
\alias{IrmsDualInletData-class}
\title{IrmsDualInletData reference class}
\description{
IrmsDualInletData reference class
}
\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}
}}
\note{
not sure yet what should go in here
}

