% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IrmsDualInletDataClass.R
\docType{class}
\name{IrmsDualInletData}
\alias{IrmsDualInletData}
\title{IrmsDualInletData reference class}
\arguments{
\item{summarize}{whether to show whole data table or just the summary
ggplot data}
}
\description{
IrmsDualInletData reference class
}
\section{Methods}{

\describe{
\item{\code{convert_data(value, data_type)}}{function converts data table entries to their appropriate data types - overwrite in derived classes for more specialized behaviour}

\item{\code{get_data_table(...)}}{retrieve the data table}

\item{\code{get_mass_data(masses = names(.self$plotOptions$masses), melt = FALSE, ...)}}{get the mass trace data for specific masses, can be provided in \code{melt = TRUE} long format
for easy use in ggplot style plotting}

\item{\code{init_irms_data()}}{initialize irms data container}

\item{\code{make_ggplot(...)}}{generate a ggplot object for the data in this IrmsData object}

\item{\code{plot(...)}}{plot data with standard plot functions (fast) to standard output}
}}
\note{
not implemented yet for any actual data reading
}
\seealso{
\link{IrmsData}, \link{IrmsContinuousFlowData}
initialize irms data container
by default, returns all data table columns that are enabled with show = TRUE
}

