% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{IsodatClumpedCO2File}
\alias{IsodatClumpedCO2File}
\title{Clumped dual inlet data class}
\description{
Clumped dual inlet data class
}
\seealso{
\link{IrmsDualInletData}, \link{BinaryFile}, \link{IsodatFile}, \link{IsodatDualInletFile}
}

