% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{IsodatDualInletFile}
\alias{IsodatDualInletFile}
\title{Clumped dual inlet data class}
\description{
Clumped dual inlet data class
}
\section{Methods}{

\describe{
\item{\code{initialize(file, ...)}}{initialize BinaryFile object, requires a file path}

\item{\code{process(...)}}{process the raw data to fill the data list}
}}
\seealso{
\link{BinaryFile}, \link{IsodatFile}, \link{IrmsDualInletData}, \link{IrmsData}
}

