% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/IsodatHydrogenContinuousFlowFileClass.R
\docType{class}
\name{IsodatHydrogenContinuousFlowFile}
\alias{IsodatHydrogenContinuousFlowFile}
\title{H-CSIA DataClass}
\description{
Objects of this class hold the isotopic data from compound specific hydrogen
isotope analysis recorded in Isodat file formats (currently supported isodat
version is 2.0 for chromatographic and peak table data and isodat version 2.5 and
3.0 for chromatographic data only).
}
\details{
This class is derived from \link{IrmsContinuousFlowData} which defines a number
of useful plotting, export and data access methods. This class also derived
\link{BinaryFile} which provides functionality for interacting with the
underlying \link{IsodatFile}.
}
\section{Methods}{

\describe{
\item{\code{cleanup(clean_raw = TRUE, clean_keys = TRUE, ...)}}{clean up the object by removing the raw data and keys (and other large but only transiently important information) from memory}

\item{\code{get_info(show = c())}}{Get basic information about the object}

\item{\code{initialize(file, ...)}}{initialize BinaryFile object, requires a file path}

\item{\code{load(...)}}{load the data from the file and generate key lookup}

\item{\code{plot_data(y, ylab = "", title = "data peaks")}}{plot the data of the actual sample peaks, see \code{plot_peak_table} for details on syntax}

\item{\code{plot_refs(y, ylab = "", title = "references")}}{plot the data of the reference peaks, see \code{plot_peak_table} for details on syntax}

\item{\code{process(...)}}{process the raw data to fill the data list}

\item{\code{reevaluate_peak_table()}}{reevalutes the peak table (not currently implemented)}
}}
\seealso{
\link{BinaryFile}, \link{IsodatFile}, \link{IrmsContinuousFlowData}, \link{IrmsData}
}

