\docType{package}
\name{isoread-package}
\alias{isoread-package}
\alias{isoreadinfo}
\title{isoread package}
\description{
R interface to IRMS (isotope ratio mass spectrometry) file
formats typically used in stable isotope geochemistry.
}
\details{
See \code{\link{isoread}} for details on how to use.
}
\author{
Sebastian Kopf \email{seb.kopf
}
\keyword{package}

