% Generated by roxygen2 (4.0.0): do not edit by hand
\name{isoread}
\alias{isoread}
\title{Read isotope data files}
\usage{
isoread(files, type, ...)
}
\arguments{
\item{file}{path to the file(s) to read}

\item{type}{type of the files to be read
\itemize{
   \item{'H_CSIA'}{ = compound specific IRMS data for hydrogen isotopes}
}}

\item{...}{parameters passed to the \code{load} and \code{process} functions of the IsodatFile objects}
}
\value{
List of file \code{type} specific objects.
\itemize{
   \item{'H_CSIA'}{ = instance(s) of \code{\link{IsodatHydrogenContinuousFlowFile}} which implements \code{\link{IrmsContinuousFlowData}}}.
}
If file names start with a number,
then the number is used as key in the list, otherwise the whole filename is the key.
If there is only one file, the object is returned directly.
}
\description{
Reads isodat file(s) and returns the contents as file type specific
instances of \code{\link{BinaryFile}} / \code{\link{IrmsDataClass}} (extends both).
}

