% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{isoread_folder}
\alias{isoread_folder}
\title{Reads all isodat files in a folder.}
\usage{
isoread_folder(folder, type, extension = ".cf", ...)
}
\description{
See \code{\link{isoread}} for paramter and return value details.
}

